Move chemistry and agnesi into ACEradials
