+++ lib/ACEradials/Project.toml
@@ -5,10 +5,9 @@ version = "0.1.0"
 [deps]
 ACEbase = "14bae519-eb20-449c-a949-9c58ed33163e"
 ChainRulesCore = "d360d2e6-b24c-11e9-a2a3-2a2ae2dbcce4"
-# EquivariantTensors is not used yet; it is declared as a dependency for the
-# upcoming converters (to_TransSelSplines, orthonormal_Rnl_basis).
 EquivariantTensors = "5e107534-7145-4f8f-b06f-47a52840c895"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 Lux = "b2108857-7c20-44ae-9111-449ecde12c47"
 LuxCore = "bb33d45b-7691-41d6-9220-0943567d0623"
 Polynomials4ML = "03c4bcba-a943-47e9-bfa1-b1661fc2974f"
@@ -16,11 +15,20 @@ Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 WithAlloc = "fb1aa66a-603c-4c1d-9bc4-66947c7b08dd"
 
+[weakdeps]
+AtomsBase = "a963bdd2-2df7-4f54-a1ee-49d51e6be12a"
+
+[extensions]
+AtomsBaseExt = "AtomsBase"
+
 [compat]
 ACEbase = "0.4.5"
+AtomsBase = "0.5"
 ChainRulesCore = "1.25.1"
+DecoratedParticles = "0.1.4"
 EquivariantTensors = "0.5"
 ForwardDiff = "0.10, 1"
+LinearAlgebra = "1.11"
 Lux = "1"
 LuxCore = "1.2.4"
 Polynomials4ML = "0.5.6"
@@ -30,10 +38,12 @@ WithAlloc = "0.1.1"
 julia = "1.11"
 
 [extras]
+AtomsBase = "a963bdd2-2df7-4f54-a1ee-49d51e6be12a"
+DecoratedParticles = "023d0394-cb16-4d2d-a5c7-724bed42bbb6"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Zygote = "e88e6eb3-aa80-5325-afca-941959d7151f"
 
 [targets]
-test = ["Test", "Zygote", "StaticArrays", "Random"]
+test = ["Test", "Zygote", "StaticArrays", "Random", "AtomsBase", "DecoratedParticles"]
