+++ lib/ACEradials/test/runtests.jl
@@ -3,4 +3,5 @@ using Test
 
 @testset "ACEradials.jl" begin
     @testset "Learnable & Splined Rnl" begin include("test_radials.jl"); end
+    @testset "Agnesi transform" begin include("test_agnesi.jl"); end
 end
